pyproject.toml
[tool.poetry]
name = "djauthtest"
version = "0.1.0"
description = ""
authors = ["bogdankozlovsky <bogdan.k@datarails.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
dj-rest-auth = "^2.2.3"
django-allauth = "^0.49.0"
PyYAML = "^6.0"
uritemplate = "^4.1.1"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
